+++ src/pallet/action/package.clj
@@ -60,7 +60,6 @@
        (aptitude
         install -q -y
         ~(if (:allow-untrusted opts) "--allow-untrusted" "")
-        ~(string/join " " (map #(str "-t " %) (:enable opts)))
         ~(string/join
           " "
           (for [[action packages] (group-by :action packages)
@@ -110,7 +109,6 @@
       (stevedore/script
        (apt-get
         -q -y install
-        ~(string/join " " (map #(str "-t " %) (:enable opts)))
         ~(string/join
           " "
           (for [[action packages] (group-by :action packages)
@@ -415,7 +413,7 @@
    "installonlypkgs=%s %s" (string/join " " packages) default-installonlypkgs))
 
 (defmethod configure-package-manager :aptitude
-  [session packager {:keys [priority prox] :or {priority 50} :as options}]
+  [session packager {:keys [priority proxy] :or {priority 50} :as options}]
   (remote-file*
    session
    (format "/etc/apt/apt.conf.d/%spallet" priority)
@@ -427,7 +425,7 @@
    :literal true))
 
 (defmethod configure-package-manager :apt
-  [session packager {:as options}]
+  [session packager options]
   (configure-package-manager session :aptitude options))
 
 (defmethod configure-package-manager :yum