+++ src/pallet/action_plan.clj
@@ -609,11 +609,12 @@
 (defn phase-for-target
   "Return the phase for the target"
   [session]
-  (let [phase (:phase session)]
+  (let [phase-key (:phase session)
+        env-phase (-> session :environment :phases phase-key)]
     (or
-     (phase (-> session :server :phases))
-     (phase (:inline-phases session))
-     (phase (-> session :group :phases)))))
+     (phase-key (-> session :server :phases))
+     (phase-key (:inline-phases session))
+     (phase-key (-> session :group :phases)))))
 
 (defn build-for-target
   "Create the action plan by calling the current phase for the target group."