+++ src/pallet/main_invoker.clj
@@ -9,7 +9,9 @@
    [pallet.configure :as configure]
    [pallet.environment :as environment]
    [pallet.utils :as utils]
-   [pallet.main :as main]))
+   [pallet.main :as main])
+  (:use
+   [pallet.task :only [exit-task-exception]]))
 
 (defn log-info
   [admin-user]
@@ -87,7 +89,7 @@
                            (re-find #"provider .* not configured" msg))
                         (binding [*out* *err*]
                           (println msg)
-                          (throw pallet.main/exit-task-exception))
+                          (throw exit-task-exception))
                         (throw e)))))]
     (if compute
       (try