+++ test/pallet/live_test.clj
@@ -46,6 +46,8 @@
   image-lists
   {:amzn-linux [{:os-family :amzn-linux :os-64-bit true}]
    :aws-ubuntu-10-10 [{:os-family :ubuntu :image-id "ami-08f40561"}]
+   :aws-ubuntu-12-04 [{:os-family :ubuntu :image-id "ami-3c994355"
+                       :os-64-bit true}]
    ;; individual images from default-images
    :ubuntu-lucid [{:os-family :ubuntu :os-version-matches "10.04"
                    :os-64-bit false}]