+++ test/pallet/main_test.clj
@@ -7,16 +7,6 @@
    clojure.test
    pallet.test-utils))
 
-(deftest parse-as-qualified-symbol-test
-  (is (nil? (parse-as-qualified-symbol "a")))
-  (is (nil? (parse-as-qualified-symbol "a.b")))
-  (is (= ['a.b 'a.b/c] (parse-as-qualified-symbol "a.b/c"))))
-
-(deftest map-and-resolve-symbols-test
-  (is (= {'pallet.main-test-ns/x 1}
-         (reduce map-and-resolve-symbols {}
-                 ["a" "b" "1" "pallet.main-test-ns/x"])))
-  (is (= 1 (var-get (find-var 'pallet.main-test-ns/x)))))
 
 (defmacro with-err-str
   "Evaluates exprs in a context in which *err* is bound to a fresh