+++ test/pallet/task/providers_test.clj
@@ -5,15 +5,6 @@
    clojure.test
    pallet.test-utils))
 
-
-(with-private-vars [pallet.task.providers
-                    [provider-properties enabled?]]
-
-  (deftest provider-properties-test
-    (is (map? (provider-properties)))
-    (is (every? string? (keys (provider-properties))))
-    (is (every? string? (vals (provider-properties)))))
-
-  (deftest enabled?-test
-    (is (enabled? "java.lang.String"))
-    (is (not (enabled? "java.lang.StringThatDoesntExist")))))
+(deftest providers-output-test
+  (let [out (with-out-str (providers))]
+    (is (.contains out "node-list"))))